fix: publish caveman runtime packaging
