+++ profiles/juliusbrussee_caveman/for-forgecat/CLAUDE.md
@@ -1,5 +1,15 @@
 # CLAUDE.md — caveman
 
+## Forgecat packaging note
+
+This converted Forgecat profile does not install the upstream top-level
+`bin/`, `src/`, or full `hooks/` helper tree directly. Runtime helper files
+that hooks and commands need are packaged under `skills/caveman/...`, because
+Forgecat installs skill directories across Claude Code, Codex, and Cursor.
+Top-level `hooks/caveman-activate.js` and
+`hooks/caveman-mode-tracker.js` are launchers that dispatch to the installed
+skill helper path.
+
 ## README is a product artifact
 
 README = product front door. Non-technical people read it to decide if caveman worth install. Treat like UI copy.
