+++ profiles/juliusbrussee_caveman/for-forgecat/hooks/caveman-mode-tracker.js
@@ -1,133 +1,42 @@
 #!/usr/bin/env node
-// caveman — UserPromptSubmit hook to track which caveman mode is active
-// Inspects user input for /caveman commands and writes mode to flag file
+// Forgecat hook launcher.
+// The real runtime lives under the installed caveman skill so its support
+// files are installed by Forgecat's normal skill-resource copy.
 
-const fs = require('fs');
-const path = require('path');
-const os = require('os');
-const { execFileSync } = require('child_process');
-const { getDefaultMode, safeWriteFlag, readFlag, VALID_MODES } = require('./caveman-config');
+(async () => {
+  const fs = await import('node:fs');
+  const path = await import('node:path');
+  const processMod = await import('node:process');
+  const { spawnSync } = await import('node:child_process');
 
-// Modes handled by their own slash commands (/caveman-commit, etc.) — not
-// selectable via /caveman <arg>.
-const INDEPENDENT_MODES = new Set(['commit', 'review', 'compress']);
+  const process = processMod.default;
+  const cwd = process.cwd();
+  const candidates = [
+    path.join(cwd, '.claude', 'skills', 'caveman', 'hooks', 'caveman-mode-tracker.js'),
+    path.join(cwd, '.cursor', 'skills', 'caveman', 'hooks', 'caveman-mode-tracker.js'),
+    path.join(cwd, '.agents', 'skills', 'caveman', 'hooks', 'caveman-mode-tracker.js'),
+  ];
+  const target = candidates.find((p) => fs.existsSync(p));
 
-const claudeDir = process.env.CLAUDE_CONFIG_DIR || path.join(os.homedir(), '.claude');
-const flagPath = path.join(claudeDir, '.caveman-active');
-
-let input = '';
-process.stdin.on('data', chunk => { input += chunk; });
-process.stdin.on('end', () => {
-  try {
-    const data = JSON.parse(input);
-    const prompt = (data.prompt || '').trim().toLowerCase();
-
-    // Natural language activation (e.g. "activate caveman", "turn on caveman mode",
-    // "talk like caveman"). README tells users they can say these, but the hook
-    // only matched /caveman commands — flag file and statusline stayed out of sync.
-    if (/\b(activate|enable|turn on|start|talk like)\b.*\bcaveman\b/i.test(prompt) ||
-        /\bcaveman\b.*\b(mode|activate|enable|turn on|start)\b/i.test(prompt)) {
-      if (!/\b(stop|disable|turn off|deactivate)\b/i.test(prompt)) {
-        const mode = getDefaultMode();
-        if (mode !== 'off') {
-          safeWriteFlag(flagPath, mode);
-        }
-      }
-    }
-
-    // /caveman-stats [--share] — block the prompt and inject stats output as
-    // the hook's reason. The script reads the active session log, so we pass
-    // transcript_path through when Claude Code provides it.
-    const statsMatch = /^\/caveman(?::caveman)?-stats(?:\s+(.*))?$/.exec(prompt);
-    if (statsMatch) {
-      const tailArgs = (statsMatch[1] || '').trim().split(/\s+/).filter(Boolean);
-      try {
-        const statsPath = path.join(__dirname, 'caveman-stats.js');
-        const argv = [statsPath];
-        if (data.transcript_path) argv.push('--session-file', data.transcript_path);
-        if (tailArgs.includes('--share')) argv.push('--share');
-        if (tailArgs.includes('--all')) argv.push('--all');
-        const sinceIdx = tailArgs.indexOf('--since');
-        if (sinceIdx !== -1 && tailArgs[sinceIdx + 1]) {
-          argv.push('--since', tailArgs[sinceIdx + 1]);
-        }
-        const out = execFileSync(process.execPath, argv, { encoding: 'utf8', timeout: 5000 });
-        process.stdout.write(JSON.stringify({ decision: 'block', reason: out.trim() }));
-      } catch (e) {
-        process.stdout.write(JSON.stringify({
-          decision: 'block',
-          reason: 'caveman-stats: could not run stats script.\nTry manually: node hooks/caveman-stats.js'
-        }));
-      }
-      return;
-    }
-
-    // Match /caveman commands
-    if (prompt.startsWith('/caveman')) {
-      const parts = prompt.split(/\s+/);
-      const cmd = parts[0]; // /caveman, /caveman-commit, /caveman-review, etc.
-      const arg = parts[1] || '';
-
-      let mode = null;
-
-      if (cmd === '/caveman-commit') {
-        mode = 'commit';
-      } else if (cmd === '/caveman-review') {
-        mode = 'review';
-      } else if (cmd === '/caveman-compress' || cmd === '/caveman:caveman-compress') {
-        mode = 'compress';
-      } else if (cmd === '/caveman' || cmd === '/caveman:caveman') {
-        // Bare /caveman → activate at configured default
-        if (!arg) {
-          mode = getDefaultMode();
-        } else if (arg === 'off' || arg === 'stop' || arg === 'disable') {
-          mode = 'off';
-        } else if (arg === 'wenyan-full') {
-          // Canonical alias — config stores as 'wenyan'
-          mode = 'wenyan';
-        } else if (VALID_MODES.includes(arg) && !INDEPENDENT_MODES.has(arg)) {
-          mode = arg;
-        }
-        // Unknown arg → mode stays null, flag untouched (no silent overwrite)
-      }
-
-      if (mode && mode !== 'off') {
-        safeWriteFlag(flagPath, mode);
-      } else if (mode === 'off') {
-        try { fs.unlinkSync(flagPath); } catch (e) {}
-      }
-    }
-
-    // Detect deactivation — natural language and slash commands
-    if (/\b(stop|disable|deactivate|turn off)\b.*\bcaveman\b/i.test(prompt) ||
-        /\bcaveman\b.*\b(stop|disable|deactivate|turn off)\b/i.test(prompt) ||
-        /\bnormal mode\b/i.test(prompt)) {
-      try { fs.unlinkSync(flagPath); } catch (e) {}
-    }
-
-    // Per-turn reinforcement: emit a structured reminder when caveman is active.
-    // The SessionStart hook injects the full ruleset once, but models lose it
-    // when other plugins inject competing style instructions every turn.
-    // This keeps caveman visible in the model's attention on every user message.
-    //
-    // Skip independent modes (commit, review, compress) — they have their own
-    // skill behavior and the base caveman rules would conflict.
-    // readFlag enforces symlink-safe read + size cap + VALID_MODES whitelist.
-    // If the flag is missing, corrupted, oversized, or a symlink pointing at
-    // something like ~/.ssh/id_rsa, readFlag returns null and we emit nothing
-    // — never inject untrusted bytes into model context.
-    const activeMode = readFlag(flagPath);
-    if (activeMode && !INDEPENDENT_MODES.has(activeMode)) {
-      process.stdout.write(JSON.stringify({
-        hookSpecificOutput: {
-          hookEventName: "UserPromptSubmit",
-          additionalContext: "CAVEMAN MODE ACTIVE (" + activeMode + "). " +
-            "Drop articles/filler/pleasantries/hedging. Fragments OK. " +
-            "Code/commits/security: write normal."
-        }
-      }));
-    }
-  } catch (e) {
-    // Silent fail
+  if (!target) {
+    process.stderr.write(
+      'caveman-mode-tracker: installed runtime not found. Checked:\n' +
+      candidates.map((p) => `  - ${p}`).join('\n') + '\n'
+    );
+    process.exit(1);
   }
+
+  const input = fs.readFileSync(0);
+  const result = spawnSync(process.execPath, [target, ...process.argv.slice(2)], {
+    cwd,
+    env: process.env,
+    input,
+    encoding: 'utf8',
+  });
+  if (result.stdout) process.stdout.write(result.stdout);
+  if (result.stderr) process.stderr.write(result.stderr);
+  process.exit(result.status == null ? 1 : result.status);
+})().catch((err) => {
+  process.stderr.write(`caveman-mode-tracker: ${err && err.stack ? err.stack : err}\n`);
+  process.exit(1);
 });
