+++ profiles/juliusbrussee_caveman/for-forgecat/skills/caveman/hooks/caveman-mode-tracker.js
@@ -1,0 +1,114 @@
+#!/usr/bin/env node
+// caveman — UserPromptSubmit hook to track which caveman mode is active
+// Inspects user input for /caveman commands and writes mode to flag file
+
+const fs = require('fs');
+const path = require('path');
+const os = require('os');
+const { execFileSync } = require('child_process');
+const { getDefaultMode, safeWriteFlag, readFlag, VALID_MODES } = require('./caveman-config');
+
+const INDEPENDENT_MODES = new Set(['commit', 'review', 'compress']);
+
+const claudeDir = process.env.CLAUDE_CONFIG_DIR || path.join(os.homedir(), '.claude');
+const flagPath = path.join(claudeDir, '.caveman-active');
+
+let input = '';
+process.stdin.on('data', chunk => { input += chunk; });
+process.stdin.on('end', () => {
+  try {
+    const data = JSON.parse(input);
+    const prompt = (data.prompt || '').trim().toLowerCase();
+
+    if (/\b(activate|enable|turn on|start|talk like)\b.*\bcaveman\b/i.test(prompt) ||
+        /\bcaveman\b.*\b(mode|activate|enable|turn on|start)\b/i.test(prompt)) {
+      if (!/\b(stop|disable|turn off|deactivate)\b/i.test(prompt)) {
+        const mode = getDefaultMode();
+        if (mode !== 'off') {
+          safeWriteFlag(flagPath, mode);
+        }
+      }
+    }
+
+    const statsMatch = /^\/caveman(?::caveman)?-stats(?:\s+(.*))?$/.exec(prompt);
+    if (statsMatch) {
+      const tailArgs = (statsMatch[1] || '').trim().split(/\s+/).filter(Boolean);
+      try {
+        const statsPath = path.join(__dirname, 'caveman-stats.js');
+        const argv = [statsPath];
+        if (data.transcript_path) argv.push('--session-file', data.transcript_path);
+        if (tailArgs.includes('--share')) argv.push('--share');
+        if (tailArgs.includes('--all')) argv.push('--all');
+        const sinceIdx = tailArgs.indexOf('--since');
+        if (sinceIdx !== -1 && tailArgs[sinceIdx + 1]) {
+          argv.push('--since', tailArgs[sinceIdx + 1]);
+        }
+        const out = execFileSync(process.execPath, argv, {
+          encoding: 'utf8',
+          timeout: 5000,
+          stdio: ['ignore', 'pipe', 'pipe'],
+        });
+        process.stdout.write(JSON.stringify({ decision: 'block', reason: out.trim() }));
+      } catch (e) {
+        const detail = e && e.stderr ? String(e.stderr).trim() : '';
+        process.stdout.write(JSON.stringify({
+          decision: 'block',
+          reason: detail || 'caveman-stats: could not run stats script.\nTry manually: node hooks/caveman-stats.js'
+        }));
+      }
+      return;
+    }
+
+    if (prompt.startsWith('/caveman')) {
+      const parts = prompt.split(/\s+/);
+      const cmd = parts[0];
+      const arg = parts[1] || '';
+
+      let mode = null;
+
+      if (cmd === '/caveman-commit') {
+        mode = 'commit';
+      } else if (cmd === '/caveman-review') {
+        mode = 'review';
+      } else if (cmd === '/caveman-compress' || cmd === '/caveman:caveman-compress') {
+        mode = 'compress';
+      } else if (cmd === '/caveman' || cmd === '/caveman:caveman') {
+        if (!arg) {
+          mode = getDefaultMode();
+        } else if (arg === 'off' || arg === 'stop' || arg === 'disable') {
+          mode = 'off';
+        } else if (arg === 'wenyan-full') {
+          mode = 'wenyan';
+        } else if (VALID_MODES.includes(arg) && !INDEPENDENT_MODES.has(arg)) {
+          mode = arg;
+        }
+      }
+
+      if (mode && mode !== 'off') {
+        safeWriteFlag(flagPath, mode);
+      } else if (mode === 'off') {
+        try { fs.unlinkSync(flagPath); } catch (e) {}
+      }
+    }
+
+    if (/\b(stop|disable|deactivate|turn off)\b.*\bcaveman\b/i.test(prompt) ||
+        /\bcaveman\b.*\b(stop|disable|deactivate|turn off)\b/i.test(prompt) ||
+        /\bnormal mode\b/i.test(prompt)) {
+      try { fs.unlinkSync(flagPath); } catch (e) {}
+    }
+
+    const activeMode = readFlag(flagPath);
+    if (activeMode && !INDEPENDENT_MODES.has(activeMode)) {
+      process.stdout.write(JSON.stringify({
+        hookSpecificOutput: {
+          hookEventName: 'UserPromptSubmit',
+          additionalContext: 'CAVEMAN MODE ACTIVE (' + activeMode + '). ' +
+            'Drop articles/filler/pleasantries/hedging. Fragments OK. ' +
+            'Code/commits/security: write normal.'
+        }
+      }));
+    }
+  } catch (e) {
+    // Silent fail
+  }
+});
