+++ profiles/juliusbrussee_caveman/for-forgecat/skills/caveman/hooks/caveman-statusline.sh
