+++ profiles/juliusbrussee_caveman/for-forgecat/skills/caveman/src/rules/caveman-openclaw-bootstrap.md
